src/core/digital_twin.py
import numpy as np
from scipy.integrate import odeint
from typing import Callable, List, Tuple, Any
from src.utils.error_handler import error_handler, logger

class DigitalTwin:
    def __init__(self, physical_system_model: Callable):
        """
        初始化数字孪生对象。

        Args:
            physical_system_model (Callable): 物理系统模型函数。
        """
        self.physical_system_model = physical_system_model
        logger.info("DigitalTwin initialized with physical system model.")

    @error_handler
    def simulate(self, initial_conditions: List[float], time_steps: List[float]) -> np.ndarray:
        """
        模拟物理系统的状态变化。

        Args:
            initial_conditions (List[float]): 初始条件。
            time_steps (List[float]): 时间步长。

        Returns:
            np.ndarray: 系统状态随时间的变化。
        """
        states = odeint(self.physical_system_model, initial_conditions, time_steps)
        logger.info(f"Simulation completed for {len(time_steps)} time steps.")
        return states

    @error_handler
    def monitor(self, sensor_data: np.ndarray) -> List[Any]:
        """
        监测系统状态，检测异常。

        Args:
            sensor_data (np.ndarray): 传感器数据。

        Returns:
            List[Any]: 检测到的异常列表。
        """
        if not hasattr(self.physical_system_model, 'detect_anomalies'):
            raise AttributeError("Physical system model does not have a 'detect_anomalies' method.")
        
        anomalies = self.physical_system_model.detect_anomalies(sensor_data)
        logger.info(f"Monitoring completed. Detected {len(anomalies)} anomalies.")
        return anomalies

    @error_handler
    def optimize(self, objective_function: Callable, constraints: List[Tuple]) -> Any:
        """
        优化系统参数。

        Args:
            objective_function (Callable): 优化目标函数。
            constraints (List[Tuple]): 优化约束条件。

        Returns:
            Any: 优化后的参数。
        """
        if not hasattr(self.physical_system_model, 'optimize'):
            raise AttributeError("Physical system model does not have an 'optimize' method.")
        
        optimal_parameters = self.physical_system_model.optimize(objective_function, constraints)
        logger.info("Optimization completed.")
        return optimal_parameters

    @error_handler
    def update_model(self, new_model: Callable) -> None:
        """
        更新物理系统模型。

        Args:
            new_model (Callable): 新的物理系统模型函数。
        """
        self.physical_system_model = new_model
        logger.info("Physical system model updated.")

    @error_handler
    def validate_model(self, validation_data: np.ndarray) -> float:
        """
        验证模型的准确性。

        Args:
            validation_data (np.ndarray): 用于验证的数据。

        Returns:
            float: 模型的准确性评分。
        """
        # 这里应该实现具体的验证逻辑
        accuracy = 0.0  # 替换为实际的准确性计算
        logger.info(f"Model validation completed. Accuracy: {accuracy}")
        return accuracy